Updated explore-nearby-universities
