+++ src/pages/nearby-university/nearby-university.css
@@ -1,0 +1,70 @@
+
+.subheading {
+    font-size: 1.5rem;
+    font-weight: bold;
+}
+
+.place-buttons .btn {
+    justify-content: center;
+    align-items: center;
+    padding: 10px 12px;
+    flex-shrink: 0;
+    border-radius: 8px;
+    border: 1.5px solid #E6E6E6;
+    background: #E6E6E6;
+    color: #000;
+    font-family: Inter, sans-serif;
+    font-size: 18px;
+    font-style: normal;
+    font-weight: 650;
+    line-height: normal;
+
+}
+
+.btn-black {
+    background-color: black!important;
+    color: white!important;
+    border: none;
+}
+
+.btn-white {
+    background-color: white;
+    color: black;
+    border: 1px solid black;
+
+}
+
+
+
+@media (max-width: 767.98px) {
+    .place-buttons {
+        flex-direction: column;
+        align-items: stretch;
+
+    }
+
+    .place-buttons .btn {
+        width: 100%;
+        font-size: 14px;
+        margin: 5px 0;
+    }
+}
+
+@media (min-width: 768px) {
+    .pad-lg-20 {
+        padding-left: 20px !important;
+        padding-right: 20px !important;
+    }
+
+}
+
+
+@media (max-width: 576px) {
+    .padx-sm-3 {
+        padding-left: 30px !important;
+        padding-right: 30px !important;
+    }
+
+}
+
+
